Todos los modelos creados

+++ db/migrate/20180929215244_create_publications.rb
@@ -6,8 +6,8 @@ class CreatePublications < ActiveRecord::Migration[5.0]
       t.datetime :startdate
       t.datetime :enddate
       t.string :place
-      t.double :latitude
-      t.double :longitude
+      t.decimal :latitude
+      t.decimal :longitude
       t.integer :type
       t.integer :user
       t.integer :parent
